// grader_labextension/src/model/assignmentSettings.ts
/**
 * Grader Extension API Schemas
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 0.1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { SubmissionPeriod } from './submissionPeriod';


export interface AssignmentSettings { 
    late_submission?: Array<SubmissionPeriod>;
}